fix: make all tests run again
